requirements.txt
affinegap==1.12
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
attrs==24.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.4
Brotli==1.1.0
BTrees==6.1
cachetools==5.5.0
categorical-distance==1.9
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
colormap==1.3.0
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
dedupe_Levenshtein_search==1.4.5
deep-translator==1.11.4
DoubleMetaphone==1.1
et_xmlfile==2.0.0
executing==2.0.1
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2025.2.0
gevent==24.11.1
geventhttpclient==2.3.3
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
googletrans-temp==2.3.0
gradio_client==1.7.0
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
haversine==2.8.1
highered==0.2.1
hpack==4.0.0
hstspreload==2024.11.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.28.1
hyperframe==6.0.1
idna==2.10
ImageHash==4.3.1
imageio==2.35.1
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kaitaistruct==0.10
keyboard==0.13.5
kiwisolver==1.4.7
lazy_loader==0.4
libretranslatepy==2.1.4
locust==2.32.5
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
MouseInfo==0.1.3
msgpack==1.1.0
multidict==6.2.0
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
networkx==3.4
numpy==2.1.0
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
paramiko==3.5.1
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
pefile==2023.2.7
persistent==6.1
pillow==11.1.0
platformdirs==4.2.2
playwright==1.55.0
prompt_toolkit==3.0.47
propcache==0.3.1
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.0.0
pure_eval==0.2.3
py4j==0.10.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydivert==2.1.0
pydub==0.25.1
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.18.0
pygsheets==2.0.6
pyhacrf-datamade==0.2.8
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyLBFGS==0.2.0.16
PyMsgBox==1.0.9
PyMuPDF==1.25.3
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.1.4
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pyspark==4.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2024.1
PyWavelets==1.8.0
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.9.4
rsa==4.9
ruff==0.9.4
safehttpx==0.1.6
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.23.1
selenium-wire==5.1.0
semantic-version==2.10.0
setuptools==75.3.0
shellingham==1.5.4
simplecosine==1.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.45.3
threadpoolctl==3.5.0
tifffile==2024.9.20
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
Wand==0.6.13
wcwidth==0.2.13
webcolors==24.11.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zope.deferredimport==5.0
zope.event==5.0
zope.index==7.0
zope.interface==7.1.1
zope.proxy==6.1
zstandard==0.23.0
